ops subsumed into blas/lapack
