+++ src/blaslapack/syev.rs
@@ -101,3 +101,38 @@ impl Syev for f32
     }
 }
 //}}}
+
+//{{{ struct: SymEigRaw
+pub(crate) struct SymEigRaw<T>
+{
+    pub eigvecs_data: Vec<T>,
+    pub eigvals:      Vec<T>,
+}
+//}}}
+//{{{ fun: symeig_raw
+/// Shared SYEV algorithm. Consumes the cloned matrix data; returns raw eigenvec/eigenval buffers.
+pub(crate) fn symeig_raw<T>(
+    mut a_data: Vec<T>,
+    n: usize,
+) -> Result<SymEigRaw<T>, Error>
+where
+    T: Syev
+        + crate::common::One
+        + crate::common::Zero
+        + crate::common::Field
+        + Default
+        + Copy
+        + super::common::AsI32,
+{
+    let mut eigvals = vec![T::zero(); n];
+
+    let mut work = vec![T::zero(); 1];
+    T::syev(b'V', b'L', n as i32, &mut a_data, n as i32, &mut eigvals, &mut work, -1)?;
+
+    let lwork = work[0].as_i32();
+    let mut work = vec![T::zero(); lwork as usize];
+    T::syev(b'V', b'L', n as i32, &mut a_data, n as i32, &mut eigvals, &mut work, lwork)?;
+
+    Ok(SymEigRaw { eigvecs_data: a_data, eigvals })
+}
+//}}}
