+++ src/dmatrix/lu.rs
@@ -12,7 +12,7 @@
 use super::DMatrix;
 use crate::blaslapack::getrf::{self, Getrf};
 use crate::common::{Field, One, Zero};
-use crate::ops::lu::lu_raw;
+use crate::blaslapack::getrf::lu_raw;
 //}}}
 //{{{ dep imports
 use thiserror::Error;
