+++ src/dmatrix/matmul.rs
@@ -12,7 +12,7 @@ use super::DMatrix;
 use crate::blaslapack::gemm::Gemm;
 use crate::blaslapack::gemv::Gemv;
 use crate::common::{Field, MatMul, One, Zero};
-use crate::ops::matmul::matmul_dispatch;
+use crate::blaslapack::gemm::matmul_dispatch;
 use crate::smatrix::SMatrix;
 //}}}
 //--------------------------------------------------------------------------------------------------
