+++ src/lib.rs
@@ -40,7 +40,6 @@
 mod blaslapack;
 mod common;
 mod expression;
-mod ops;
 mod subviews;
 //}}}
 //{{{ collection: public API
