+++ src/smatrix/symeig.rs
@@ -11,7 +11,7 @@ use super::SMatrix;
 use crate::blaslapack::common::AsI32;
 use crate::blaslapack::syev::{self, Syev};
 use crate::common::{Field, One, Zero};
-use crate::ops::symeig::symeig_raw;
+use crate::blaslapack::syev::symeig_raw;
 //}}}
 //{{{ dep imports
 use thiserror::Error;
